feat(debug): helpers for "has connected to Europa"
